// Website_UIB_dev/src/types/DetailGelombangTypes.ts
export interface DetailGelombangType {
    id: number;
    NamaPeriode: string;
    NamaGelombang: string;
    tanggal_mulai: string;
    tanggal_selesai: string;
    status: string;
    jenis_pendaftaran: string;
    jenjang_gelombang: string;
    IsBeasiswa: string;
    mulai_daftar: string;
    tanggal_ujian: string;
    tanggal_pengumuman: string;
    LokasiNama: string;
    mulai_verifikasi: string;
    akhir_verifikasi: string;
    tanggal_daftar_ulang: string;
    jam_mulai_ujian: string;
    jam_akhir_ujian: string;
    biaya_formulir: number;
}